backend/requirements.txt
fastapi
uvicorn
pydantic
openai
supabase
python-multipart
pypdf
python-dotenv
